Drop GT demo
I was originally going to show the translation alongside it to make it
easier to understand what I was doing, but it was too wide and I had to
crop it. It was pointed out that at this point it probably detracts from
my argument to show a webshit input box.